:sparkles: Refactor App.js
