+++ App.js
@@ -1,39 +1,26 @@
 import React, {useState} from 'react';
 import { StyleSheet, Text,
-         View, TextInput, Button,
-         ScrollView, FlatList
+         View, FlatList
        } from 'react-native';
+import GoalItem from './components/GoalItem';
+import GoalInput from './components/GoalInput';
 
 export default function App() {
-  const [enteredGoal, setEnteredGoal] = useState('');
   const [courseGoals, setCourseGoals] = useState([]);
 
-  const getInputGoalHandler = (goal) => {
-     setEnteredGoal(goal);
-  }
-  const addGoalHandler = () => {
+  const addGoalHandler = (goal) => {
     setCourseGoals(courseGoals => [...courseGoals,
-       {id: Math.random().toString(), value: enteredGoal}]);
+       {id: Math.random().toString(), value: goal}]);
   }
   return (
     <View style={styles.screen}>
-      <View style={styles.inputContainer}>
-       <TextInput placeholder="Course Goal" 
-       style={styles.input} 
-       onChangeText={getInputGoalHandler}
-       value={enteredGoal}
-       />
-       <Button title="ADD" onPress={addGoalHandler}/>
-      </View>
-        {courseGoals.length!=0?
+        <GoalInput
+         onAddGoal = {addGoalHandler}/>
+
+         {courseGoals.length!=0?
          <FlatList data={courseGoals}
-         renderItem={({item: {value}}) =>(
-            <View style={styles.listItem}>
-               <Text>{value}</Text>
-            </View>
-          )}
-          keyExtractor={item => item.id}
-          />
+          renderItem={itemData =><GoalItem onDelete={() => console.log('Long Pressed')} itemData = {itemData}/>}
+          keyExtractor={item => item.id}/>
         :
         <Text>Start adding your goals!</Text>
         }   
@@ -44,22 +31,5 @@ export default function App() {
 const styles = StyleSheet.create({
    screen:{
     padding: 50
-   },
-   inputContainer:{
-    flexDirection:'row',
-    justifyContent:'space-between',
-    alignItems:'center'
-   },
-   input: {
-    width:'80%',
-    borderBottomColor:'black',
-    borderBottomWidth: 1
-    },
-    listItem:{
-      padding: 10,
-      marginVertical: 5,
-      backgroundColor: '#ccc',
-      borderColor:'#f0f0',
-      borderWidth: 1
-    }
+   }
 });
